+++ python/ray/data/_internal/object_extensions/arrow.py
@@ -7,7 +7,7 @@ from packaging.version import parse as parse_version
 
 import ray.data._internal.object_extensions.pandas
 from ray._common.serialization import pickle_dumps
-from ray._private.arrow_utils import _check_pyarrow_version, get_pyarrow_version
+from ray.data.arrow_utils import _check_pyarrow_version, get_pyarrow_version
 from ray.util.annotations import PublicAPI
 
 # First, assert Arrow version is w/in expected bounds
