+++ python/ray/data/datasource/file_datasink.py
@@ -4,7 +4,6 @@ from typing import TYPE_CHECKING, Any, Dict, Iterable, Optional
 from urllib.parse import urlparse
 
 from ray._common.retry import call_with_retry
-from ray._private.arrow_utils import add_creatable_buckets_param_if_s3_uri
 from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
 from ray.data._internal.execution.interfaces import TaskContext
 from ray.data._internal.planner.plan_write_op import WRITE_UUID_KWARG_NAME
@@ -13,6 +12,7 @@ from ray.data._internal.util import (
     RetryingPyFileSystem,
     _is_local_scheme,
 )
+from ray.data.arrow_utils import add_creatable_buckets_param_if_s3_uri
 from ray.data.block import Block, BlockAccessor
 from ray.data.context import DataContext
 from ray.data.datasource.datasink import Datasink, WriteResult
