+++ python/ray/data/tests/conftest.py
@@ -11,9 +11,9 @@ import pytest
 
 import ray
 from ray._common.test_utils import wait_for_condition
-from ray._private.arrow_utils import get_pyarrow_version
 from ray._private.internal_api import get_memory_info_reply, get_state_from_address
 from ray.data._internal.tensor_extensions.arrow import ArrowTensorArray
+from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.block import BlockExecStats, BlockMetadata
 from ray.data.constants import TENSOR_COLUMN_NAME
 from ray.data.context import DEFAULT_TARGET_MAX_BLOCK_SIZE, DataContext, ShuffleStrategy
