+++ python/ray/data/tests/expressions/test_arithmetic.py
@@ -10,8 +10,8 @@ import pytest
 from packaging.version import parse as parse_version
 
 import ray
-from ray._private.arrow_utils import get_pyarrow_version
 from ray.data._internal.util import rows_same
+from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.expressions import col, lit
 from ray.data.tests.conftest import *  # noqa
 from ray.tests.conftest import *  # noqa
