+++ python/ray/data/tests/test_tensor_extension.py
@@ -8,7 +8,6 @@ import pyarrow as pa
 import pytest
 from packaging.version import parse as parse_version
 
-from ray._private.arrow_utils import get_pyarrow_version
 from ray.data import DataContext
 from ray.data._internal.tensor_extensions.arrow import (
     ArrowConversionError,
@@ -27,6 +26,7 @@ from ray.data._internal.tensor_extensions.pandas import TensorArray, TensorDtype
 from ray.data._internal.tensor_extensions.utils import (
     create_ragged_ndarray,
 )
+from ray.data.arrow_utils import get_pyarrow_version
 
 
 @pytest.mark.parametrize("tensor_format", ["v1", "v2"])
