+++ python/ray/data/tests/test_with_column.py
@@ -5,8 +5,8 @@ import pytest
 from pkg_resources import parse_version
 
 import ray
-from ray._private.arrow_utils import get_pyarrow_version
 from ray.data._internal.util import rows_same
+from ray.data.arrow_utils import get_pyarrow_version
 from ray.data.datatype import DataType
 from ray.data.exceptions import UserCodeException
 from ray.data.expressions import col, lit, udf
